Remove ERROR and Fatal Error message from Docker

+++ Customer/Gemini-Chef.sh
@@ -14,7 +14,10 @@ then
   echo "Pull Chef Server from Registry..."
   docker pull registry.gemini-systems.net/gemini/gemini-chef
   echo "Continue to run chef ..."
-  docker rm -f gemini-chef
+  if docker ps -a |grep -a gemini-chef > /dev/null; then
+        docker rm -f gemini-chef
+  fi
+
   ip=`curl -s http://ipecho.net/plain; echo`
   docker run -it -p 443:443 --privileged -v /etc/chef-server/ --name gemini-chef -h $ip -d registry.gemini-systems.net/gemini/gemini-chef
 fi
